Youtube_Director/application/requirements.txt
Flask
selenium
gunicorn
PyJWT